+++ src/app.js
@@ -13,12 +13,7 @@ export class App {
       { route: 'object-dist/publisher', moduleId: PLATFORM.moduleName('objektide_levi/publisher/publisher'), name: 'publisher' },
       { route: 'object-dist', moduleId: PLATFORM.moduleName('objektide_levi/object-dist/object-dist'), name: 'object-dist' },
       { route: 'crm/agents', moduleId: PLATFORM.moduleName('crm/agents/agents'), name: 'agents' },
-      { route: 'crm/pipeline', moduleId: PLATFORM.moduleName('./crm/pipeline/pipeline'), name: 'pipeline' },
-      { route: 'crm/opportunities-list-view', moduleId: PLATFORM.moduleName('./crm/opportunities/opportunities'), name: 'opportunities-list-view' },
-      { route: 'crm/opportunities-card-view', moduleId: PLATFORM.moduleName('./crm/opportunities-card-view/opportunities-card-view'), name: 'opportunities-card-view' },
-      { route: 'crm/opportunity-edit', moduleId: PLATFORM.moduleName('./crm/opportunity-edit/opportunity-edit'), name: 'opportunity-edit' },
-      { route: 'crm/opportunities', moduleId: PLATFORM.moduleName('./crm/opportunities-test/opportunities-test'), name: 'opportunities' },
-
+      { route: 'crm/opportunities', moduleId: PLATFORM.moduleName('./crm/opportunities/opportunities'), name: 'opportunities' },
     ]);
     this.router = router;
   }
