+++ src/crm/components/listview/listview.html
@@ -18,7 +18,8 @@
           <td>${opportunity.price}€</td>
           <td>${opportunity.stage}</td>
           <td>
-            <a route-href="route: opportunity-edit; params.bind: { opportunity: opportunity }">Edit</a> |
+            <a href="#">Edit</a> |
+            <!-- <a route-href="route: opportunity-edit; params.bind: { opportunity: opportunity }">Edit</a> | -->
             <a href="#">Delete</a>
             <!-- <a click.delegate="deleteOpportunity(opportunity.id)">Delete</a> -->
         </td>
