+++ src/crm/components/opportunity-modal/opportunity-modal.js
@@ -1,6 +1,6 @@
 import { inject } from 'aurelia-framework';
 import { EventAggregator } from 'aurelia-event-aggregator';
-import { OpportunitiesService } from '../service/opportunities-service.js';
+import { OpportunitiesService } from '../../service/opportunities-service.js';
 
 @inject(EventAggregator)
 export class OpportunityModal {
