--- src/crm/opportunities-card-view/opportunities-card-view.html
@@ -1,20 +1,0 @@
-<template>
-  <require from="crm/opportunity-modal/opportunity-modal"></require>
-
-  <div class="container mt-4">
-    <opportunity-modal></opportunity-modal>
-
-    <div class="row mt-4">
-      <div class="col-4" repeat.for="opportunity of opportunities">
-        <div class="card" style="height: 12rem; margin-bottom: 20px;">
-          <div class="card-body">
-            <h5 class="card-title">${opportunity.name}</h5>
-            <p class="card-text">${opportunity.description}</p>
-            <p class="card-text">${opportunity.stage}</p>
-            <p class="card-text"><b>${opportunity.price}€</b></p>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
