--- src/crm/opportunities-card-view/opportunities-card-view.js
@@ -1,14 +1,0 @@
-import { OpportunitiesService } from '../service/opportunities-service';
-
-export class OpportunitiesCardView {
-  constructor() {
-    this.opportunitiesService = new OpportunitiesService();
-  }
-
-  attached() {
-    this.opportunitiesService.getOpportunities()
-      .then(
-        data => this.opportunities = data
-      );
-  }
-}
