--- src/crm/opportunities-test/opportunities-test.html
@@ -1,31 +1,0 @@
-<template>
-  <require from="crm/opportunity-modal/opportunity-modal"></require>
-  <div class="container mt-4">
-    <opportunity-modal></opportunity-modal>
-
-    <div class="row mt-4 justify-content-between">
-      <div class="col-10">
-        <router-view></router-view>
-      </div>
-      <div class="col-1" style="padding-top: 15%;">
-        <div class="">
-          <div class="row">
-            <a route-href="route: listview">
-              <img src="assets/img/list.svg" alt="list">
-            </a>
-          </div>
-          <div class="row mt-3">
-            <a route-href="route: cardview">
-              <img src="assets/img/card.svg" alt="card">
-            </a>
-          </div>
-          <div class="row mt-3">
-            <a route-href="route: pipeline">
-              <img src="assets/img/pipeline.svg" alt="pipeline">
-            </a>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
