+++ src/crm/opportunities/opportunities.js
@@ -1,21 +1,16 @@
-import { OpportunitiesService } from '../service/opportunities-service';
+import { PLATFORM } from 'aurelia-pal';
 
-export class OpportunitiesList {
-  constructor() {
-    this.opportunitiesService = new OpportunitiesService();
+export class OpportunitiesTest {
+  configureRouter(config, router) {
+    config.title = 'Opportunities';
+    config.options.pushState = true;
+    // config.options.root = '/';
+    config.map([
+      { route: ['', 'pipeline'], moduleId: PLATFORM.moduleName('crm/components/pipeline/pipeline'), name: 'pipeline' },
+      { route: 'listview', moduleId: PLATFORM.moduleName('crm/components/listview/listview'), name: 'listview' },
+      { route: 'cardview', moduleId: PLATFORM.moduleName('crm/components/cardview/cardview'), name: 'cardview' },
+
+    ]);
+    this.router = router;
   }
-
-  attached() {
-    this.opportunitiesService.getOpportunities()
-      .then(
-        data => this.opportunities = data
-      );
-  }
-
-  // deleteOpportunity(id) {
-  //   this.opportunities.filter(function(obj) {
-  //     return obj.id !== id;
-  //   });
-  // }
 }
-
