--- src/crm/opportunity-modal/opportunity-modal.html
@@ -1,72 +1,0 @@
-<template>
-  <div class="row">
-
-    <div class="col">
-      <h1>Opportunities</h1>
-    </div>
-    <div class="col">
-      <div class="row float-right">
-        <div class="dropdown">
-          <img src="assets/img/favorite.svg" alt="favorite">
-          <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown"
-            aria-haspopup="true" aria-expanded="false">
-            Add new
-          </button>
-          <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-            <a class="dropdown-item" data-toggle="modal" href="#opportunityModal">Opportunity</a>
-            <a class="dropdown-item" href="#">Column</a>
-          </div>
-
-          <div class="modal fade" id="opportunityModal" tabindex="-1" role="dialog"
-            aria-labelledby="opportunityModalLabel" aria-hidden="true">
-            <div class="modal-dialog modal-dialog-centered" role="document">
-              <div class="modal-content">
-                <div class="modal-header">
-                  <h5 class="modal-title" id="opportunityModalLabel">Add new opportunity</h5>
-                  <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                    <span aria-hidden="true">&times;</span>
-                  </button>
-                </div>
-                <div class="modal-body">
-                  <form role="form" id="opportunityForm" submit.delegate="newOpportunity()">
-                    <div class="form-group">
-                      <label for="recipient-name" class="col-form-label">Name</label>
-                      <input type="text" class="form-control" id="opportunity-name" value.bind="name">
-                    </div>
-                    <div class="form-group">
-                      <label for="message-text" class="col-form-label">Description</label>
-                      <textarea class="form-control" id="opportunity-description" value.bind="description"></textarea>
-                    </div>
-                    <div class="input-group">
-                      <label for="message-text" class="col-form-label">Price</label>
-                      <div class="input-group">
-                        <input type="text" class="form-control" aria-label="Price" aria-describedby="basic-addon2"
-                          value.bind="price">
-                        <div class="input-group-append">
-                          <span class="input-group-text" id="basic-addon2">€</span>
-                        </div>
-                      </div>
-                    </div>
-                    <div class="form-group">
-                      <label for="message-text" class="col-form-label">Stage</label>
-                      <select class="form-control" value.bind="stage">
-                        <option value="new">New</option>
-                        <option value="proposition">Proposition</option>
-                        <option value="won">Won</option>
-                      </select>
-                    </div>
-                    <div class="modal-footer">
-                      <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                      <button type="button" class="btn btn-primary" click.delegate="newOpportunity(contact)"
-                        data-dismiss="modal">Save changes</button>
-                    </div>
-                  </form>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
