--- src/crm/pipeline/pipeline.html
@@ -1,86 +1,0 @@
-<template>
-  <require from="crm/opportunity-modal/opportunity-modal"></require>
-  <div class="container mt-4">
-    <opportunity-modal></opportunity-modal>
-
-    <div class="row mt-4">
-      <div class="col-11">
-        <div class="row">
-          <div class="col-4" reorderable-group-for.bind="new">
-            <div class="row-4 mb-4">
-              <h4>New</h4>
-            </div>
-            <div class="list-group-item-secondary rounded border-primary p-2 m-1"
-              reorderable-repeat.for="oppurtunity of new" reorderable-after-reordering="fruitsOrdered"
-              reorderable-group="lifes">
-              <div class="row">
-                <div class="col">
-                  <h5>${oppurtunity.name}</h5>
-                </div>
-                <div class="col text-right">
-                  <img src="assets/img/edit.svg" alt="edit">
-                  <img src="assets/img/delete.svg" alt="delete" click.delegate="deleteNewOpportunity($index)">
-                </div>
-              </div>
-              <div>${oppurtunity.description}</div>
-              <div>${oppurtunity.price}€</div>
-            </div>
-          </div>
-          <div class="col-4" reorderable-group-for.bind="proposition">
-            <div class="row-4 mb-4">
-              <h4>Proposition</h4>
-            </div>
-            <div class="list-group-item-primary rounded border-primary p-2 m-1"
-              reorderable-repeat.for="oppurtunity of proposition" reorderable-after-reordering="animalsOrdered"
-              reorderable-group="lifes">
-              <div>
-                <div class="row">
-                  <div class="col">
-                    <h5>${oppurtunity.name}</h5>
-                  </div>
-                  <div class="col text-right">
-                    <img src="assets/img/edit.svg" alt="edit">
-                    <img src="assets/img/delete.svg" alt="delete" click.delegate="deletePropositionOpportunity($index)">
-                  </div>
-                </div>
-              </div>
-              <div>${oppurtunity.description}</div>
-              <div>${oppurtunity.price}€</div>
-            </div>
-          </div>
-          <div class="col-4" reorderable-group-for.bind="won">
-            <div class="row-4 mb-4">
-              <h4>Won</h4>
-            </div>
-            <div class="list-group-item-success rounded border-primary p-2 m-1"
-              reorderable-repeat.for="oppurtunity of won" reorderable-after-reordering="insectsOrdered"
-              reorderable-group="lifes">
-              <div class="row">
-                <div class="col">
-                  <h5>${oppurtunity.name}</h5>
-                </div>
-                <div class="col text-right">
-                  <img src="assets/img/edit.svg" alt="edit">
-                  <img src="assets/img/delete.svg" alt="delete" click.delegate="deleteOpportunity($index)">
-                </div>
-              </div>
-              <div>${oppurtunity.description}</div>
-              <div>${oppurtunity.price}€</div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div class="col-1 h-100">
-        <div class="row">
-          <img src="assets/img/list.svg" alt="list">
-        </div>
-        <div class="row">
-          <img src="assets/img/card.svg" alt="card">
-        </div>
-        <div class="row">
-          <img src="assets/img/pipeline.svg" alt="pipeline">
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
