--- src/crm/pipeline/pipeline.js
@@ -1,79 +1,0 @@
-import { inject } from 'aurelia-framework';
-import { EventAggregator } from 'aurelia-event-aggregator';
-import { PipelineService } from '../service/pipeline-service';
-
-@inject(EventAggregator)
-export class Pipeline {
-  name = '';
-  description = '';
-  price = '';
-  stage = '';
-  _subscriptions = [];
-
-  constructor(ea) {
-    this.new = [];
-    this.proposition = [];
-    this.won = [];
-
-    this.pipelineService = new PipelineService();
-    this.ea = ea;
-
-    ea.subscribe('reorderable-group:intention-changed', intention => {
-      this.intention = intention;
-    });
-
-    // ea.subscribe('newOpportunityCreation',
-    //   description => this.eventNewOpportunityCreation(description));
-  }
-
-  // eventNewOpportunityCreation(description) {
-  // }
-
-  attached() {
-    this.pipelineService.getNewOpportunities()
-      .then(
-        data => this.new = data
-      );
-    this.pipelineService.getPropositionOpportunities()
-      .then(
-        data => this.proposition = data
-      );
-    this.pipelineService.getWonOpportunities()
-      .then(
-        data => this.won = data
-      );
-  }
-
-  deleteWonOpportunity(index) {
-    this.won.splice(index, 1);
-  }
-
-  deleteNewOpportunity(index) {
-    this.new.splice(index, 1);
-  }
-
-  deletePropositionOpportunity(index) {
-    this.proposition.splice(index, 1);
-  }
-
-  objArrayReordered(objArray, change) {
-    /* eslint no-console: 0 */
-    console.log(JSON.stringify(objArray));
-    console.log('change', change);
-  }
-
-  fruitsOrdered(list, change) {
-    console.log('newFruits: ' + list);
-    console.log('change', change);
-  }
-
-  animalsOrdered(list, change) {
-    console.log('animals: ' + list);
-    console.log('change', change);
-  }
-
-  insectsOrdered(list, change) {
-    console.log('insects: ' + list);
-    console.log('change', change);
-  }
-}
